text_to_multimedia_ai_pipeline/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
mlx
Pillow
python-multipart
pytest
opencv-python
